+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * feat: Add `click` and refactor argparse.
 * feat: Add `lai` command, as a shortcut to `lambda-ai`.
 * feat: Add `lai ls` for displaying instances in a table.
+* feat: Add `lai images` for displaying images in a table.
 
 # 1.2.0
 * docs: Set development status to production.